+++ tests/testthat/test-points.R
@@ -3,17 +3,17 @@ library(testthat)
 context("points")
 
 ## TODO: Rename context
-library(rangl)
+library(anglr)
 library(maptools)
 data(wrld_simpl)
 pts <- as(as(wrld_simpl, "SpatialLinesDataFrame"), "SpatialPointsDataFrame")
 mpts <- as(as(wrld_simpl, "SpatialLinesDataFrame"), "SpatialMultiPointsDataFrame")
 test_that("points works", {
-  expect_silent(rangl(pts))
+  expect_silent(anglr(pts))
 })
 
 
 test_that("multi-points works", {
-  expect_that(rangl(mpts), is_a("pointmesh"))
-  expect_that(rangl(geometry(mpts)), is_a("pointmesh"))
+  expect_that(anglr(mpts), is_a("pointmesh"))
+  expect_that(anglr(geometry(mpts)), is_a("pointmesh"))
 })